requirements.txt
confluent-kafka==2.3.0
Flask==3.0.2
mysql-connector-python==8.3.0